requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-easy-select2==1.5.7
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-graphql-social-auth==0.1.4
django-imagekit==4.0.2
django-import-export==2.3.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==8.1.0
promise==2.3
pycparser==2.20
PyJWT==2.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
